fix(clippy): rustc 1.96 lint sweep — workspace clippy-clean again
~30 sites across kevy-store / kevy-lua / kevy-lua-host / kevy-cli /
kevy-chaos / kevy-rt / kevy-embedded / kevy-madvise / kevy-uring:
auto-deref, RangeInclusive::contains, collapsible if, len==0,
io::Error::other, doc lazy-continuations (leading '+' prose read as
list items), Vec<Box<Value>> → Vec<Value> on the bio-drop channel
(boxing was redundant — the Vec already heap-owns), type aliases for
scan-page returns + hash_pairs, sha1 w-schedule enumerate, LPOS
take(scan_limit) instead of a manual counter.

Semantics-neutral throughout; workspace 122 test suites green.
Unblocks the aarch64 CI Clippy step (red since the runner's stable
moved to 1.96).

+++ crates/kevy-store/src/small_hash.rs
@@ -137,7 +137,7 @@ impl SmallHashData {
             let new_used = new_used_i as usize;
             // Shift the tail to make room (or close the gap).
             if delta != 0 {
-                self.buf.copy_within(old_end..used, (val_off + new_vlen) as usize);
+                self.buf.copy_within(old_end..used, val_off + new_vlen);
                 if delta < 0 {
                     // Zero freed tail.
                     self.buf[new_used..used].fill(0);
